+++ src/main/java/frc/robot/RobotContainer.java
@@ -7,7 +7,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.RobotMap.mapControllers;
 import frc.robot.commands.UpdatePoseEstimator;
-import frc.robot.commands.Drive.DriveSimple;
+import frc.robot.commands.Drive.Simple;
 import frc.robot.subsystems.Drivetrain;
 import frc.robot.subsystems.Vision;
 
@@ -20,7 +20,7 @@ public class RobotContainer {
   public RobotContainer() {
 
     subDrivetrain.setDefaultCommand(
-        new DriveSimple(subDrivetrain, conDriver, true, true));
+        new Simple(subDrivetrain, conDriver, true, true));
 
     subVision.setDefaultCommand(new UpdatePoseEstimator(subDrivetrain, subVision));
